refactor for new version
