+++ client.go
@@ -1,0 +1,73 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package agollo
+
+import (
+	"github.com/zouyx/agollo/v4/agcache/memory"
+	"github.com/zouyx/agollo/v4/cluster/roundrobin"
+	"github.com/zouyx/agollo/v4/component"
+	"github.com/zouyx/agollo/v4/component/log"
+	"github.com/zouyx/agollo/v4/component/notify"
+	"github.com/zouyx/agollo/v4/component/serverlist"
+	"github.com/zouyx/agollo/v4/env"
+	"github.com/zouyx/agollo/v4/env/config"
+	jsonFile "github.com/zouyx/agollo/v4/env/file/json"
+	"github.com/zouyx/agollo/v4/extension"
+	"github.com/zouyx/agollo/v4/protocol/auth/sign"
+	"github.com/zouyx/agollo/v4/storage"
+)
+
+type Client struct {
+	initAppConfigFunc func() (*config.AppConfig, error)
+}
+
+func Create() *Client {
+	extension.SetCacheFactory(&memory.DefaultCacheFactory{})
+	extension.SetLoadBalance(&roundrobin.RoundRobin{})
+	extension.SetFileHandler(&jsonFile.FileHandler{})
+	extension.SetHTTPAuth(&sign.AuthSignature{})
+	storage.InitConfigCache()
+	return &Client{}
+}
+
+func (c *Client) Start() error {
+	return c.StartWithConfig(nil)
+}
+
+func (c *Client) StartWithConfig(loadAppConfig func() (*config.AppConfig, error)) error {
+	// 有了配置之后才能进行初始化
+	if err := env.InitConfig(loadAppConfig); err != nil {
+		return err
+	}
+
+	notify.InitAllNotifications(nil)
+	serverlist.InitSyncServerIPList()
+
+	//first sync
+	if err := notify.SyncConfigs(); err != nil {
+		return err
+	}
+	log.Debug("init notifySyncConfigServices finished")
+
+	//start long poll sync config
+	go component.StartRefreshConfig(&notify.ConfigComponent{})
+
+	log.Info("agollo start finished ! ")
+
+	return nil
+}
