+++ cluster/roundrobin/round_robin_test.go
@@ -21,7 +21,7 @@ import (
 	"testing"
 
 	. "github.com/tevid/gohamcrest"
-	"github.com/zouyx/agollo/v3/env"
+	"github.com/zouyx/agollo/v4/env"
 )
 
 const servicesConfigResponseStr = `[{
