+++ component/common.go
@@ -21,9 +21,9 @@ import (
 	"fmt"
 	"net/url"
 
-	"github.com/zouyx/agollo/v3/env"
-	"github.com/zouyx/agollo/v3/env/config"
-	"github.com/zouyx/agollo/v3/utils"
+	"github.com/zouyx/agollo/v4/env"
+	"github.com/zouyx/agollo/v4/env/config"
+	"github.com/zouyx/agollo/v4/utils"
 )
 
 //AbsComponent 定时组件
