+++ component/notify/componet_notify_test.go
@@ -27,12 +27,12 @@ import (
 	"time"
 
 	. "github.com/tevid/gohamcrest"
-	_ "github.com/zouyx/agollo/v3/cluster/roundrobin"
-	"github.com/zouyx/agollo/v3/env"
-	"github.com/zouyx/agollo/v3/env/config"
-	jsonConfig "github.com/zouyx/agollo/v3/env/config/json"
-	_ "github.com/zouyx/agollo/v3/env/file/json"
-	"github.com/zouyx/agollo/v3/extension"
+	_ "github.com/zouyx/agollo/v4/cluster/roundrobin"
+	"github.com/zouyx/agollo/v4/env"
+	"github.com/zouyx/agollo/v4/env/config"
+	jsonConfig "github.com/zouyx/agollo/v4/env/config/json"
+	_ "github.com/zouyx/agollo/v4/env/file/json"
+	"github.com/zouyx/agollo/v4/extension"
 )
 
 var (
