+++ extension/cache.go
@@ -17,18 +17,18 @@
 
 package extension
 
-import "github.com/zouyx/agollo/v3/agcache"
+import "github.com/zouyx/agollo/v4/agcache"
 
 var (
-	gobalCacheFactory agcache.CacheFactory
+	globalCacheFactory agcache.CacheFactory
 )
 
 //GetCacheFactory 获取CacheFactory
 func GetCacheFactory() agcache.CacheFactory {
-	return gobalCacheFactory
+	return globalCacheFactory
 }
 
 //SetCacheFactory 替换CacheFactory
 func SetCacheFactory(cacheFactory agcache.CacheFactory) {
-	gobalCacheFactory = cacheFactory
+	globalCacheFactory = cacheFactory
 }
