+++ extension/cache_test.go
@@ -23,7 +23,7 @@ import (
 	"testing"
 
 	. "github.com/tevid/gohamcrest"
-	"github.com/zouyx/agollo/v3/agcache"
+	"github.com/zouyx/agollo/v4/agcache"
 )
 
 type TestCacheFactory struct {
