+++ extension/load_balance_test.go
@@ -22,7 +22,7 @@ import (
 	"testing"
 
 	. "github.com/tevid/gohamcrest"
-	"github.com/zouyx/agollo/v3/env/config"
+	"github.com/zouyx/agollo/v4/env/config"
 )
 
 type TestLoadBalance struct {
