+++ mock_server_test.go
@@ -23,7 +23,7 @@ import (
 	"net/http/httptest"
 	"strings"
 
-	"github.com/zouyx/agollo/v3/env/config"
+	"github.com/zouyx/agollo/v4/env/config"
 )
 
 const (
