+++ repository.go
@@ -20,11 +20,11 @@ package agollo
 import (
 	"strconv"
 
-	"github.com/zouyx/agollo/v3/agcache"
-	"github.com/zouyx/agollo/v3/component/log"
-	"github.com/zouyx/agollo/v3/component/notify"
-	"github.com/zouyx/agollo/v3/storage"
-	"github.com/zouyx/agollo/v3/utils"
+	"github.com/zouyx/agollo/v4/agcache"
+	"github.com/zouyx/agollo/v4/component/log"
+	"github.com/zouyx/agollo/v4/component/notify"
+	"github.com/zouyx/agollo/v4/storage"
+	"github.com/zouyx/agollo/v4/utils"
 )
 
 //GetConfig 根据namespace获取apollo配置
