+++ start.go
@@ -19,40 +19,17 @@ package agollo
 
 import (
 	"container/list"
-	"github.com/zouyx/agollo/v3/agcache"
-	_ "github.com/zouyx/agollo/v3/agcache/memory"
-	"github.com/zouyx/agollo/v3/cluster"
-	_ "github.com/zouyx/agollo/v3/cluster/roundrobin"
-	"github.com/zouyx/agollo/v3/component"
-	"github.com/zouyx/agollo/v3/component/log"
-	"github.com/zouyx/agollo/v3/component/notify"
-	"github.com/zouyx/agollo/v3/component/serverlist"
-	"github.com/zouyx/agollo/v3/env"
-	"github.com/zouyx/agollo/v3/env/config"
-	"github.com/zouyx/agollo/v3/env/file"
-	_ "github.com/zouyx/agollo/v3/env/file/json"
-	"github.com/zouyx/agollo/v3/extension"
-	"github.com/zouyx/agollo/v3/protocol/auth"
-	_ "github.com/zouyx/agollo/v3/protocol/auth/sign"
-	"github.com/zouyx/agollo/v3/storage"
-	_ "github.com/zouyx/agollo/v3/utils/parse/normal"
-	_ "github.com/zouyx/agollo/v3/utils/parse/properties"
-	_ "github.com/zouyx/agollo/v3/utils/parse/yml"
+	"github.com/zouyx/agollo/v4/agcache"
+	"github.com/zouyx/agollo/v4/cluster"
+	"github.com/zouyx/agollo/v4/component/log"
+	"github.com/zouyx/agollo/v4/env/file"
+	"github.com/zouyx/agollo/v4/extension"
+	"github.com/zouyx/agollo/v4/protocol/auth"
+	"github.com/zouyx/agollo/v4/storage"
+	_ "github.com/zouyx/agollo/v4/utils/parse/normal"
+	_ "github.com/zouyx/agollo/v4/utils/parse/properties"
+	_ "github.com/zouyx/agollo/v4/utils/parse/yml"
 )
-
-var (
-	initAppConfigFunc func() (*config.AppConfig, error)
-)
-
-//InitCustomConfig init config by custom
-func InitCustomConfig(loadAppConfig func() (*config.AppConfig, error)) {
-	initAppConfigFunc = loadAppConfig
-}
-
-//start apollo
-func Start() error {
-	return startAgollo()
-}
 
 //SetSignature 设置自定义 http 授权控件
 func SetSignature(auth auth.HTTPAuth) {
@@ -102,33 +79,10 @@ func AddChangeListener(listener storage.ChangeListener) {
 
 //RemoveChangeListener 增加变更监控
 func RemoveChangeListener(listener storage.ChangeListener) {
-	storage.AddChangeListener(listener)
+	storage.RemoveChangeListener(listener)
 }
 
 //GetChangeListeners 获取配置修改监听器列表
 func GetChangeListeners() *list.List {
 	return storage.GetChangeListeners()
-}
-
-func startAgollo() error {
-	// 有了配置之后才能进行初始化
-	if err := env.InitConfig(initAppConfigFunc); err != nil {
-		return err
-	}
-
-	notify.InitAllNotifications(nil)
-	serverlist.InitSyncServerIPList()
-
-	//first sync
-	if err := notify.SyncConfigs(); err != nil {
-		return err
-	}
-	log.Debug("init notifySyncConfigServices finished")
-
-	//start long poll sync config
-	go component.StartRefreshConfig(&notify.ConfigComponent{})
-
-	log.Info("agollo start finished ! ")
-
-	return nil
 }
