+++ utils/parse/yaml/parser_test.go
@@ -18,8 +18,8 @@
 package yaml
 
 import (
-	"github.com/zouyx/agollo/v3/utils"
-	"github.com/zouyx/agollo/v3/utils/parse"
+	"github.com/zouyx/agollo/v4/utils"
+	"github.com/zouyx/agollo/v4/utils/parse"
 	"testing"
 
 	. "github.com/tevid/gohamcrest"
@@ -62,4 +62,3 @@ func TestYAMLParserOnException(t *testing.T) {
 	m := convertToMap(nil)
 	Assert(t, m, NilVal())
 }
-
